sales-insight-automator/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.3.3
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
aiosmtplib==3.0.1       # async SMTP — new in Phase 3
